Build Spydate for x86, so a 32-bit program can have a 32-bit debugger
  dotnet build src\Spydate.App -p:PlatformTarget=x86

produces Spydate-x86.exe beside the 64-bit one, the way dnSpy ships two.
It is a build and not a switch because several things in the interop
layer are decided by the width of the debugger process itself, and none
of them can be a runtime flag.

Thread contexts. A 32-bit thread has the same CONTEXT whoever reads it,
but a 64-bit debugger reaches it through Wow64GetThreadContext and a
32-bit one through plain GetThreadContext - the Wow64 pair exists to let
a 64-bit process reach across, and calling it on your own kind returns
nothing. Wow64ThreadContext is now X86ThreadContext and takes which pair
to use. A 64-bit target from the 32-bit build is refused with a reason.

Pointer-sized arguments. ReadProcessMemory, WriteProcessMemory,
VirtualProtectEx and FlushInstructionCache all took the address as
ulong, which is pointer-width on x64 by coincidence and corrupts the
stack on x86. This was not theoretical: it was the first thing the
32-bit build did, dying in ReadProcessMemory with 0xC0000005.

DEBUG_EVENT. Every offset in it depends on the debugger's width: the
union starts at 12 on x86 and 16 on x64, and inside it every handle and
address is pointer-width, so each field after the first moves too. Read
at the wrong width this does not fail - it yields plausible, wrong
addresses - so the offsets are now pairs.

Register names. The mixed-mode prestub dance asked every thread for rsp
and rax by name. On a 32-bit debuggee those are not there, and a name
that is not there is not an error: the lookup returns zero, and what was
going to be done with the stack pointer is done with address zero. That
is how it failed without ever saying so. The names and the width of a
return address now follow the debuggee, and the first-call catch is off
for a 32-bit one, where it cannot work as written - PreStubWorker is
__stdcall on x86 and takes its arguments on the stack, so there is no
register to read the MethodDesc from. Cold breakpoints still arrive
through PlantPending on a later call.

Verified: the 32-bit build debugging a 32-bit native DLL under its host
reaches the loader break, rebases Mingus.dll from 0x10000000 to
0x5BA60000, arms four breakpoints and stops on one. And on a 32-bit
.NET assembly a managed breakpoint is planted and hit:
MingusHarness.Program.Report, token 0x06000046, planted at 0x56E44A0
as soon as the JIT produced it, then stopped on three times running,
with eip on the breakpoint each time.

The access violation that looked like part of this is not: the test
program crashes identically with no debugger attached at all, exit code
0xC0000005 in MingusHarness.Mingus.getCurrentLicenseCode, a P/Invoke of
its own. It runs further under the debugger (60 lines of output) than
without it (56), so nothing here provoked it.

91 tests pass on x64, which is where all of this also runs. Both builds
keep their own bin and obj, set in Directory.Build.props for every
project rather than only the app: sharing them let an x86 build quietly
replace the x64 artifacts, apphosts included, and the next x64 test run
failed 34 tests with errors naming assemblies and bitness rather than
the build that had overwritten them.

+++ src/Spydate.Debugger/Managed/ManagedDebugSession.cs
@@ -1030,7 +1030,7 @@ public sealed class ManagedDebugSession : IDisposable, IManagedEvents
                     return $"patch at {where} not applied: the process gave no handle to write through";
                 }
 
-                if (!Native.VirtualProtectEx(handle, address, (nuint)length, Native.PageReadWrite, out uint previous))
+                if (!Native.VirtualProtectEx(handle, (nuint)address, (nuint)length, Native.PageReadWrite, out uint previous))
                 {
                     return $"patch at {where} not applied: the IL page could not be made writable "
                            + $"(Win32 0x{Marshal.GetLastWin32Error():X})";
@@ -1045,13 +1045,13 @@ public sealed class ManagedDebugSession : IDisposable, IManagedEvents
                     {
                         fixed (byte* p = bytes)
                         {
-                            wrote = Native.WriteProcessMemory(handle, address, p, (nuint)length, out written);
+                            wrote = Native.WriteProcessMemory(handle, (nuint)address, p, (nuint)length, out written);
                         }
                     }
                 }
                 finally
                 {
-                    Native.VirtualProtectEx(handle, address, (nuint)length, previous, out _);
+                    Native.VirtualProtectEx(handle, (nuint)address, (nuint)length, previous, out _);
                 }
 
                 if (!wrote || (int)written != length)
@@ -1059,7 +1059,7 @@ public sealed class ManagedDebugSession : IDisposable, IManagedEvents
                     return $"patch at {where} not applied: the write failed (Win32 0x{Marshal.GetLastWin32Error():X})";
                 }
 
-                Native.FlushInstructionCache(handle, address, (nuint)length);
+                Native.FlushInstructionCache(handle, (nuint)address, (nuint)length);
                 Note($"patched {where}: {length} byte(s) written before the method was compiled");
                 return null;
             }
